# nam_viet_erp_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/clinic/appointments": {
            "post": {
                "description": "Lễ tân tạo lịch hẹn khám bệnh cho bệnh nhân",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinic"
                ],
                "summary": "Đặt lịch khám",
                "parameters": [
                    {
                        "description": "Book appointment request payload",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BookAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Appointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/clinic/appointments/{id}/check-in": {
            "post": {
                "description": "Lễ tân xác nhận bệnh nhân đã tới và đưa vào hàng đợi chờ khám",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinic"
                ],
                "summary": "Bệnh nhân Check-in",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Check-in request payload",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.CheckInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/crm/loyalty/earn": {
            "post": {
                "description": "Cộng điểm cho khách hàng dựa trên giá trị đơn hàng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRM"
                ],
                "summary": "Tích điểm thành viên",
                "parameters": [
                    {
                        "description": "Earn loyalty request payload",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.EarnLoyaltyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/finance/transactions": {
            "post": {
                "description": "Sinh phiếu thu chi, tự động chặn nếu vi phạm luật thuế, chẻ phiếu nếu lệch tiền",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance"
                ],
                "summary": "Tạo giao dịch thu/chi",
                "parameters": [
                    {
                        "description": "Create transaction request payload",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/finance/vat-allocation": {
            "post": {
                "description": "Sử dụng thuật toán Knapsack để gom nhóm các Order Items khớp với tổng giá trị hóa đơn yêu cầu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance"
                ],
                "summary": "Phân bổ hóa đơn VAT",
                "parameters": [
                    {
                        "description": "VAT allocation request payload",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VATAllocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Appointment": {
            "type": "object",
            "properties": {
                "appointment_time": {
                    "type": "string"
                },
                "check_in_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "integer"
                },
                "service_type": {
                    "description": "e.g., 'examination'",
                    "type": "string"
                },
                "status": {
                    "description": "'pending', 'waiting', 'completed', 'cancelled'",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.BookAppointmentRequest": {
            "type": "object",
            "required": [
                "appointment_time",
                "doctor_id",
                "patient_id",
                "service_type"
            ],
            "properties": {
                "appointment_time": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "integer"
                },
                "patient_id": {
                    "type": "integer"
                },
                "service_type": {
                    "type": "string"
                }
            }
        },
        "domain.CheckInRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "domain.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "flow",
                "fund_account_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "flow": {
                    "type": "string"
                },
                "fund_account_id": {
                    "type": "integer"
                },
                "ref_id": {
                    "type": "string"
                },
                "ref_type": {
                    "type": "string"
                }
            }
        },
...
        "domain.EarnLoyaltyRequest": {
            "type": "object",
            "required": [
                "amount",
                "customer_id",
                "order_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "customer_id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.VATAllocationRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "target_amount"
            ],
            "properties": {
                "customer_id": {
                    "type": "integer"
                },
                "target_amount": {
                    "type": "number"
                }
            }
        },
...
    }
}